move decorator, adjust and ignore some types
