+++ src/innov8/components/intra_sector.py
@@ -4,7 +4,7 @@ import pandas as pd
 from dash import dash_table, dcc, html
 from dash.dependencies import Input, Output
 
-from innov8.components.decorators import callback, data_access
+from innov8.decorators.data_access import callback, data_access
 
 
 # Store intermediate values
