+++ src/innov8/components/main_carousel.py
@@ -2,7 +2,7 @@ import dash_trich_components as dtc
 from dash import html
 from dash.dependencies import Input, Output
 
-from innov8.components.decorators import callback, data_access
+from innov8.decorators.data_access import callback, data_access
 
 change_query = """
 WITH growth AS (
